% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ssb_plot_shocks}
\alias{ssb_plot_shocks}
\title{Exposure-concentration (Lorenz) plot}
\usage{
ssb_plot_shocks(x, title = NULL, ...)
}
\arguments{
\item{x}{An [ssb_shock_summary()] (`ssb_shocks`) object.}

\item{title}{Optional plot title.}

\item{...}{Unused.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Lorenz curve of the shock exposure weights from [ssb_shock_summary()]: the
further the curve bows below the 45-degree line, the more the identifying
variation is concentrated in a few shocks. The effective number of shocks and
the HHI are shown in the subtitle.
}
